Make it work with program state in different file
